DAQ-1991 Plot the driver profile in the plan wizard
Also added a few more profiles in the dummy experiment service for a 
richer demo

+++ uk.ac.diamond.daq.experiment.api/src/uk/ac/diamond/daq/experiment/api/driver/DriverProfileSection.java
@@ -15,7 +15,7 @@ public class DriverProfileSection implements EditableWithListWidget {
 	
 	public DriverProfileSection() {}
 	
-	public DriverProfileSection(int start, int stop, double duration) {
+	public DriverProfileSection(double start, double stop, double duration) {
 		this.start = start;
 		this.stop = stop;
 		this.duration = duration;
